Add a .desktop file and make the setup tool work with Gnome
Previously, the setup tool could be called from the command
line, for example:

    /usr/libexec/ibus-setup-table cangjie3
    /usr/libexec/ibus-setup-table wubi-jidian86
    ...

But it could not be called from “gnome-control-center region”,
not setup icon did appear for the ibus-table engines.

To make this work, a desktop files

    /usr/share/applications/ibus-setup-{...}.desktop

need to be added where {...} is the engine name up to the first “:”
character. When the engine names are just the table names
(“cangjie3”, “wubi-jidian86”, ...), each table would need to
supply its own .desktop file

    /usr/share/applications/ibus-setup-cangjie3.desktop
    /usr/share/applications/ibus-setup-wubi-jidian86.desktop
    ...

That is not nice, especially as many tables are in different rpm packages.
It is better to have just a single .desktop file

    /usr/share/applications/ibus-setup-table.desktop

This can be used for all engines if the ibus engine names are changed like:

    cangjie3 → table:cangjie3
    wubi-jidian86 → table:wubi-jidian86
    ...

because only the part before the first “:” in the ibus engine name
is used by gnome to decide which desktop file to use.

The desktop file contains a line like:

    Exec=/usr/libexec/ibus-setup-table

and the setup tool is is called by Gnome without any command line
arguments. But Gnome sets the enviromnent variable IBUS_ENGINE_NAME
to the engine name when calling the setup tool.

So I changed the setup tool to work like this:

    IBUS_ENGINE_NAME=table:cangjie3 /usr/libexec/ibus-setup-table

or

    /usr/libexec/ibus-setup-table --engine-name table:cangjie

(The command line option is preferred over the environment variable
when both are given).