Make usage of engine name and dconf key consistent
The setup tool did write to a different dconf key than the
engine.

Make this consistent. If the file name of the table is

    /usr/share/ibus-table/tables/wubi-jidian86.db

use

   table:wubi-jidian

as the ibus engine name (listed by “ibus list-engine” and used
in the environment variable IBUS_ENGINE_NAME).

As the dconf key, use:

    /desktop/ibus/engine/table/wubi-jidian86

The name found in the source of the table (i.e. in wubi-jidian86.txt
in the source package of ibus-table-chinese), which looks like

    ### The default name of this table
    NAME = WuBi-Jidian (JiShuag) 86

should not be used for the dconf key or the engine name, it should
only be used for display in the user interface.