If keychar is empty, do not treat it as valid input
This makes typing the right shift key toggle between pinyin mode and
table mode, as intended.

In Python

    >>> '' in 'abc'
    True

Therefore, “keychar in self._valid_input_char” is True even if
keychar is empty. But if keychar is empty, one should not treat it
as valid input.

This problem occured when pressing Shift-Right, it was treated
as valid input and the following traceback was seen:

Traceback (most recent call last):
  File "/local/mfabian/src/ibus-table/engine/table.py", line 1596, in do_process_key_event
    result = self._process_key_event (key)
  File "/local/mfabian/src/ibus-table/engine/table.py", line 1618, in _process_key_event
    return self._table_mode_process_key_event (key)
  File "/local/mfabian/src/ibus-table/engine/table.py", line 1856, in _table_mode_process_key_event
    res = self._editor.add_input ( keychar )
  File "/local/mfabian/src/ibus-table/engine/table.py", line 398, in add_input
    res = self.update_candidates ()
  File "/local/mfabian/src/ibus-table/engine/table.py", line 804, in update_candidates
    self.pop_input ()
  File "/local/mfabian/src/ibus-table/engine/table.py", line 409, in pop_input
    self._tabkey_list.pop()
IndexError: pop from empty list