Change the design of the database
The old database used columns like

    id|mlen|clen|m0|m1|m2|m3|m4|category|phrase|freq|user_freq

where “m0”, “m1”, “m2”, “m3”, “m4” did hold individual
characters of the input string which was therefore limited to a fixed
maximum (could be different for each table). In this case the maximum
is 5 characters. “mlen” was the length of the input string actually
used, for example if the input string was “adlt”, then “mlen = 3”,
“m0 = a”, “m1 = d”, “m2 = l”, “m3 = t”, and “m4 = None”.
“clen” was the length of the Chinese phrase matching this input
string, for example, if “adlt” is the input string for the Chinese
phrase “苦力”, then “clen = 2”. “category” was a bitmask
indicating whether that Chinese phrase is simplified Chinese,
traditional Chinese, used in both variants of Chinese, or is a mixture
of simplified and traditional Chinese. Non-Chinese tables lacked that
“category” column.

I.e. in this example the database row in system database for the old
design would have looked like:

    345|4|2|a|d|l|t|None|1|苦力|3600000|0

(“345” is an abitrary row-id)

The new database now uses only:

    id|tabkeys|phrase|freq|user_freq

“tabkeys” is the combination of “m0” ... “m4”, it is of type TEXT
and has no artificial limit (except the limit imposed by sqlite, which
is huge). So the row in the above example would look like this in
the new database design:

    345|adlt|苦力|3600000|0

“mlen” and “clen” have been removed, they just duplicate
information, the lengths can be calculated easily when
needed. “category” has been removed as well, it can be calculated
fast and easy at runtime, there is no need to save this at build time
into the system database. Omitting “category” also makes the columns
used in Chinese and non-Chinese databases identical, avoiding lots of
special cases for Chinese and non-Chinese in the code.