Fix cursor movements in the strings committed to preëdit
Rewrite most of the code for cursor movements in the preëdit because
it did not work at all.

The old non-working code tried to use wrap around when the cursor
reached the ends of the preëdit. I think that is confusing
and implemented it without wrap around.

When there are strings committed to preëdit and a space is typed
while the current input is empty, the preëdit should be committed
(this was broken).