Fix remove_cand_from_userdb() and rename it to remove_candidate_from_user_database()
After the database redesign, tabsqlitedb.remove_phrase() takes
named arguments “tabkeys” and “phrase” instead of a complete candidate
as an array.

Also, instead of deleting a candidate from the user database only
if it is user defined and thus not in the system database (candidates
which have freq=-1 in the user database are user defined), delete
the candidate always, if it is in the user database at all.

This way, the user can not only use this function to remove user defined
candidates but also to lower the priority of candidates which are
already in the system database and have aquired a high priority by
being used. Removing these candidates from the user database lowers
their priority again.