Rename self._chars_valid to self._chars_valid_when_update_candidates_was_last_called
Long name, but this makes the purpose of that variable clear.
It is used *only* to make update_candidates() return fast when
nothing has really changed.

If the users input characters have not changed, usually no update_candidates()
is necessary, but there are some exceptional cases for example when
deleting a candidate from the user database when the candidate
list changes while the user input has not changed. In these exceptional
cases, an update of the lookup table needs to be forced.