self._strings and self._u_chars should always have the same length
As self._strings and self._u_chars *always* should have the same
length, there is no need to check “if self._strings” before
self._strings.pop() if self._u_chars is already known to be not
empty. Of course the extra check doesn’t hurt, but if self._strings
and self._u_chars ever happen to have a different length, I want it to
fail and then figure out how this could happen.