over_input() should not clear self._u_chars[]
It does this only to cause a weird side effect in
get_preedit_strings().  But that is quite nonsensical, get_predit_strings()
and _update_preedit() should be rewritten not to depend on this.

Clearing only self._u_chars[] but not self._strings[] makes these
two arrays have different lengths, but they should always have
the same length.

if clearing self._u_chars[] is removed from over_input(),
over_input() becomes useless as it only calls clear_input() now
and it can be removed.