Handle start characters correctly
Some tables (currently the “latex” table is the only one I know of)
have special “start characters”. For example, in the latex table,
the start character is “\”, i.e. every key in that table starts with
a “\”, if the first character typed is anything else but “\”, no
match is possible and therefore the typed character should be directly
committed or directly passed to the application without attempting
a match in the table.