Fix find_zi_code, it should return a list of possible tabkeys for a phrase
The documentation “Check word freq and user_freq” was very misleading.

Actually it should return a list of possible tabkeys for a phrase.
For example, if “phrase” is “你” and the table is wubi-jidian.86.txt,
the result will be ['wq', 'wqi', 'wqiy'] because that table
contains the following 3 lines matching that phrase exactly:

wq	你	597727619
wqi	你	1490000000
wqiy	你	1490000000