Do not hide the auxiliary text when there are no candidates, it might contain useful information
The auxiliary text may contain useful information even when there are
no candidates.

For example, when committing a few characters to
preëdit to define a user defined phrase, the auxiliary text shows
what key sequence will be defined for that phrase.
For example, when wubi-jidian86 is used and the phrase

    “王不各”

is commited to preëdit and no further input is currently there, the
auxiliary text shows

    “#: ggtk”

Here “ggtk” is the key sequence which will be defined for that phrase,
i.e. in future one can type “ggtk” to input “王不各”.

The auxiliary text also shows the table keys if the last character
committed to preëdit has been typed using pinyin. For example,
if one uses wubi-jidian86 and has forgotten the table codes for

    你

one can type it using pinyin by switching to pinyin mode (with the menu
or the right shift key), type “ni + Shift + 3” to get 你, commit it
to preëdit by selecting it with Control+Number in the lookup table,
then the wubi-jidian86 table codes for 你 are displayed in the auxiliary
text:

    wq wqi wqiy