Committing with the mouse or by typing Control+Number should not define new abbreviations
I avoided defining new abbreviations when committing by typing
“space” already, see:

    commit 56796c39c92697d4fcd233e42ebe8f208f0e1324
    Author: Mike FABIAN <mfabian@redhat.com>
    Date:   Wed May 7 14:07:56 2014 +0200

        Don’t define new abbreviations automatically

        For example, the latex table system database contains:

            \alpha|α|0|0

        If one types “\alp” and then selects the candidate “α”
        from the lookup table, do *not* automatically create
        (or update) an abbreviated entry

            \alp|α|0|1

        in the user database, instead create (or update) an entry using the
        full length of the tabkeys:

            \alpha|α|0|1

        Creating shortcuts automatically like this is useful in
        ibus-typing-booster but it seems useless and confusing in ibus-table.

but it didn’t work when committing using Control+Number because
the code for Control+Number got the input key sequence *before*
committing to preedit and at this point it may not yet be the complete
input key sequence of a candidate. During commit to preëdit, the
really typed input key sequence is replaced by the input key sequence
of the selected candidate, which may be longer.

Committing with the mouse inherited this error from the code copied
from committing with Control+Number.