auto_commit_to_preedit() is not needed
There is no significant difference between auto_commit_to_preedit()
and commit_to_preedit().

The only difference seems to be that auto_commit_to_preedit() tries to
commit even when self._chars_valid is empty and even when
self._candidates is empty. But in that case it will just fail with an
“index out of range” error. To catch this, auto_commit_to_preedit()
uses a “try: ... except: ...”  instead of checking whether
self._candidates is empty like commit_to_preedit() does, that seems
to be the only real difference.