When auto_commit is used and there is only one candidate, commit only when the typed characters match the tabkeys of the candidate completely
Much better for non-Chinese tables like “latex”, “ipa-x-sampa”, ...

For example, consider that “latex” table contains the following
lines matching “\al”:

    \alpha α
    \aleph ℵ

And the default for the “latex” table is to use auto_commit.
That means that with the old code, after typing “\alp”,
only the single candidate

    \alpha α

matched and was committed immediately. That is annoying because this
makes it necessary to look at the screen. It is easy to remember that
“\alpha” means α in LaTeX and if these are commited only when typed
completely, one does not have to look at the screen to type α. It is
much harder to remember how many characters of \alpha one has to type
to get a unique match. Of course committing at “\alp” already saves
two key strokes. But even with the new code one can save a few
keystrokes if one looks at the screen: Just type “\al”, notice that
α is the first candidate, then commit it with space.

The new code also avoids the distinction between Chinese/non-Chinese here:
In a Chinese table, when

    len(self._editor._chars_valid) == self._max_key_length

*and* there is only one candidate, it follows that the tabkeys
for this candidate have been typed completely. That means
one can just as well do the same check as for non-Chinese tables,
i.e.

   self._editor._chars_valid == self._editor._candidates[0][0]

instead of checking whether the maximum key length is reached.