Add options to setup for fullwidth and halfwidth mode for direct input mode
There are such options for “Table input” already but not for
“Direct input”.  But even for “Direct input”, when the fullwidth
or halfwidth mode is changed using the property menu, the change is
saved to the config and remembered, just as for “Table input”. There
seems to be no reason for this is kind of “asymmetric” behaviour.
So I added these options for “Direct mode” to the setup as well.