Make the setup tool get the default from the database for settings which are not yet set in dconf
The setup tool gets the the values for the options to
show when it starts up from the following 3 places now:

   1) dconf setting for this table
   2) database of this table
   3) Python dictionary OPTION_DEFAULTS in main.py of setup.

So it prefers dconf settings, then the database, and finally the
hardcoded values in the setup tool itself.

Before this change it did fall back immediately to the hardcoded
values OPTION_DEFAULTS if a setting was not yet set in dconf.
I.e. it showed exactly the same defaults for all tables.

But the defaults in the databases of the tables are usually there for
a reason and they should of course be presented to the user before he
first changes something in the setup tool.