Make the option in the setup tool to use “space” as a commit or page down key work
And some more improvements in the handling of commit and page up/down keys.

In future we might want to introduce separate options for commit keys
and page up/down keys in the setup tool instead of the option for the
space key behaviour. But this commit makes the existing option work
which is enough for the moment.