If the dconf key “lookuptablepagesize” changes to an invalid value, reset it to a valid value
If the user sets the dconf key “lookuptablepagesize” to a value < 1
or > len(self._editor._select_keys), set it to a valid value again in
the callback handler for dconf key changes.

After this change, if the user tries to move the slider for the lookup
table page size in the setup tool to 10 for a table like “stroke5”
which has only 9 select keys (SELECT_KEYS = 1,2,3,4,5,6,7,8,9), the
dconf key is set to 9 again and the slider moves back to 9.