Do not check whether the name in “NAME = ...” is a valid gconf key
It is not used as a gconf key anyway. We use the basename of the table
file as the gconf key, i.e. for

    /usr/share/ibus-table/tables/wubi-jidian86.db

the gconf key is “wubi-jidian86”. The gconf key should be more or
less constant, it should not change often. If it changes during an
update, the old user settings may be lost.

   NAME =

in the table source is more suitable for display to the user,
here one can write a nice, user readable string.

In the lastest master of ibus-table-chinese, wubi-jidian86.txt contains:

    NAME = WuBi-Jidian (JiShuag) 86
    NAME.zh_CN = 极点五笔（极爽字库）86
    NAME.zh_TW = 極點五筆（極爽字庫）86
    NAME.zh_HK = 極點五筆（極爽字庫）86

which made tabcreatedb.py fail because it checked whether
NAME is a valid gconf key. But we do not need that check.