Add a “setup” property
Apparently that was planned already, there was a comment which
said “Need implementation”.

I also think it is a good idea, especially under Gnome3 the setup for
the input methods is quite difficult to reach, one has to start the
“gnome-control-centre” first, then go to the “region” part, then
to the input method, then click on the cogwheel icon. That is quite
inconvenient if the setup is used regularly.

If one wanted to reduce the number of entries in the properties, I
think it would be better to remove other properties which are probably
not used all the time and are available in the setup anyway.  I guess
the properties “Phrase mode/Single character mode”, “Direct commit
mode/Normal commit mode” and “Display candidates/Hide candidates”
are probably very rarely used and it might be good enough if they are
accessible only in the setup tool.