Support prompt characters
Some tables, e.g. cangjie5.txt contain prompt character definitions:

BEGIN_CHAR_PROMPTS_DEFINITION
a 日
b 月
c 金
...

Read these when creating a database and save it to the database.

When typing, show these prompts instead of showing the typed Latin
characters in the auxiliary text, the preëdit, or in the candidate
list for the remaining keys to type to complete a match.