Add the Unicode code point of the first character of the phrase as a last ditch sort key for the candidates
In some cases, where there are several matches for *exactly* the same
input sequences *and* the system frequencies are also the same
for example

    bmr	同	1000
    bmr	𦚈	1000

in cangjie5.txt, instead of getting the order randomly, use the
Unicode code point as a last ditch sort key. Some defined order is
better than random. On top of that, this puts characters with code
points above the BMP *after* characters with code points in the
BMP. Usually this is the right thing to do because the characters from
the BMP are usually more common. Should this not be the case and the
character above the BMP really is the more common one, the system
frequency should be added to reflect this.

This does not always help: In cangjie5.txt there are also cases where
*exactly* the same input sequences matches different characters which
are all in the BMP, for example:

    bmso	豚	1000
    bmso	冢	1000
    bmso	䐁	1000

    dhe	皮	1000
    dhe	板	1000

But sorting the characters in the BMP to the top if all other sort
keys are the same is better than nothing.