Add wildcard support also to pinyin mode
The wildcards used are the same as for table mode. I.e. if
the table defines:

    MULTI_WILDCARD_CHAR = *
    SINGLE_WILDCARD_CHAR = ?

the same wildcards are used for pinyin as well.

The AUTO_WILDCARD in the table is ignored for pinyin mode though, it
is always treated as AUTO_WILDCARD = TRUE for pinyin mode, FALSE
does not make any sense for pinyin mode.