Insert a special candidate for the wildcard character itself if only a wildcard character has been typed
Resolves: rhbz#1110325 - Unable to input question marks in Wubi-jidian
See: https://bugzilla.redhat.com/show_bug.cgi?id=1110325

The same problem has been also reported here:

    https://code.google.com/p/ibus/issues/detail?id=1720

If only a wildcard character has been typed, insert a special
candidate at the first position for the wildcard character itself. For
example, if “?” is used as a wildcard character and this is the only
character typed, add a candidate ('?', '?', 0, 1000000000) in
halfwidth mode or a candidate ('?', '？', 0, 1000000000) in fullwidth
mode.  This is needed to make it possible to input the wildcard
characters themselves, if “?” acted only as a wildcard it would be
impossible to input a fullwidth question mark.