Make wildchard characters configurable
The options AUTO_WILDCARD, SINGLE_WILDCARD_CHAR, and
MULTI_WILDCARD_CHAR are now not only configurable in the table sources
but also in the setup tool and via dconf.

The default for AUTO_WILDCARD if neither specified in the table source
nor in the setup tool nor in dconf is TRUE, this is best for most
tables.

The defaults for SINGLE_WILDCARD_CHAR and MULTI_WILDCARD_CHAR
are the empty strings, this means no wildcards are used.

In some table sources default values for SINGLE_WILDCARD_CHAR and
MULTI_WILDCARD_CHAR are defined, in most cases these are

SINGLE_WILDCARD_CHAR = ?
MULTI_WILDCARD_CHAR = *