Fix prompts for array30 table and don’t use prompts in pinyin mode
The array 30 table has prompts which are longer than one character.
Like this:

BEGIN_CHAR_PROMPTS_DEFINITION
q 1^
w 2^
e 3^
...
m 7v
, 8v
. 9v
/ 0v
END_CHAR_PROMPTS_DEFINITION

This did not work with my old code.

Also, do not use the prompts in pinyin mode. This commit

    commit c9f6a2bfe833da6f6a3d67546a89e0cefd886d83
    Author: Mike FABIAN <mfabian@redhat.com>
    Date:   Wed Jun 4 12:45:19 2014 +0200

	Don’t show the prompt characters defined in the table in pinyin mode in the auxiliary text

already avoided using the prompts in pinyin mode in the auxiliary text.
But prompt characters may also appear in the candidate list to show
which characters are needed to complete a candidate (“remaining_tabkeys”)
and also in the preëdit when no candidate can be found for the typed
characters.  Fix this for pinyin mode for these cases as well.