Require Python >= 3.3
When porting to Python3, I originally kept it working with Python2 as
well. But I didn’t test with Python2 for a while and the current code
would not work anymore with Python2.  It doesn’t seem to make sense
to invest effort to keep Python2 compatibility. Better require Python >= 3.3.