When a leading invalid character is passed through, it needs to be remembered in self._prev_char
When a string is committed using commit_string(), commit_string()
remembers its last characters in self._prev_char. But when a invalid
leading character is passed through directly to the application by
returning False in the key event processing, that character needs to
be remembered as well in self._prev_char.

For example, when halfwidth letter mode but fullwidth punctuation mode
is used, _convert_to_full_width() does not convert a “.”  to “。”
if the previous character was a halfwidth digit. Probably the purpose
of this hack is to make it easier to enter floating point number like
“3.141”. A digit like “3” is an invalid input character for most
Chinese tables and thus committed or passed through to the application
immediately.  If the next character typed is “.”,
_convert_to_full_width() checks whether the previous character was a
halfwidth digit and if yet prevents conversion of “.” to ”。” to
avoid getting “3。141”. Therefore, no matter whether a character
is committed or passed through, the last character inserted into
the application needs to be remembered.