Show the table code in the candidate list when pinyin mode is used
Resolves: rhbz#1318109 - [RFE] Please show Wubi table code when in pinyin mode
See: https://bugzilla.redhat.com/show_bug.cgi?id=1318109, https://github.com/ibus/ibus/issues/1846

When using pinyin mode for a table like Wubi or Cangjie,
the reason is probably because one does not know the
Wubi or Cangjie code. So get that code from the table
and display it as well to help the user learn that code.