Add some debug code to see whether _phrases_cache is handled correctly
It needs to be reset when changing the Chinese mode. And stuff committed
to the user database needs to be removed.

Can be tested using the SmartCangJie6 table (scj6.txt) because this table uses
both

    USER_CAN_DEFINE_PHRASE = FALSE

and

    DYNAMIC_ADJUST = FALSE

so no tabkeys are ever removed individually from _phrases_cache but
_phrases_cache is reset to empty when the Chinese mode is changed (or
when other settings influencing the caching like onechar mode or
wildcard characters are changed).