Fix problems with property menus
Property menus were behaving strangely sometimes.  For example, when
choosing pinyin mode in the property menu, then moving the focus to a
client window and typing in pinyin, then opening the pinyin property
menu again, it showed table mode, not pinyin mode. Although the real
mode was pinyin mode.

Fix the property menu handling by making the code as similar as
possible with the property menu handling in ibus-typing-booster. The
property menu handling code in ibus-typing-booster code has several
bugs fixed, and it is easier to maintain.