Make key bindings configurable
Only via the command line for the moment, not yet easy to do for
normal users. I have to rewrite the setup tool eventually to make that
possible.

Resolves: https://github.com/ibus/ibus/issues/2241

This rewrite of the handling of the key bindings also fixes the
problem reported in https://github.com/ibus/ibus/issues/2241
that the Shift_L key cannot switch between Chinese and English mode
when using google-chrome.