Change the default keybinding for commit_to_preedit to ['Control+Shift_L', 'Control+Shift_R']
Related: https://github.com/kaio/ibus-table/issues/68

The previous default had a conflict with toggle_input_mode_on_off
and toggle_pinyin_mode:

    commit_to_preedit               ['Shift_R', 'Shift_L']
    toogle_input_mode_on_off        ['Shift_L']
    toggle_pinyin_mode              ['Shift_R']

This conflict didn’t matter as the toggling of the modes was only done
when the preedit was empty and commit_to_preedit can only be done
when the preedit is not empty. So there was no real conflict in practice.
But now when toggle_input_mode_on_off is also done when the preedit is
not empty to fix https://github.com/kaio/ibus-table/issues/68, then
it is not possible to use the same keybindings for these commands.

commit_to_preedit is probably used much more rarely than toggling the
input or the pinyin modes. So it is probably better to leave the
default keybindinds for toggling the input and pinyin modes as they
always were and change the default keybinding of commit_to_preedit.