Generate png versions of the ibus-table icon from the svg master
To be able to display an icon in gnome-software, svg does not work there.

I tried to use “convert” from ImageMagick to do the conversion like this:

for i in 16 22 32 48 64 128 256
do
    mkdir ${i}x${i}
    convert -background none -density 1024 -resize ${i}x${i} ibus-table.svg ${i}x${i}/ibus-table.png
    mogrify -density 96 ${i}x${i}/ibus-table.png
done

But that just produces completely empty png files. Adding the -verbose option
shows that inkscape is used for the conversion. But opening the ibus-table.svg
in inkscape just shows a white box.

So I used gimp to do the conversion.